[Bronze V] Title: 이상한 기호, Time: 80 ms, Memory: 11500 KB -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -10,8 +10,8 @@ public class Main {
 		StringBuilder sb = new StringBuilder();
 
 		StringTokenizer st = new StringTokenizer(br.readLine());
-		int A = Integer.parseInt(st.nextToken());
-		int B = Integer.parseInt(st.nextToken());
+		long A = Long.parseLong(st.nextToken());
+		long B = Long.parseLong(st.nextToken());
 		
 		System.out.println((A+B)*(A-B));
 	}
